refactor: JsonUtils
